refactor and lint
